Added integration between fitter and plotter
